+++ frontend/src/views/components/menubar/MenuBar.css
@@ -17,7 +17,7 @@
 
   position: absolute;
   top: -20px;
-  right: 10%;
+  right: 8%;
   width: 50px;
   height: 50px;
   box-shadow: 0px 5px 20px 0px rgba(80, 80, 80, 10%);
