+++ frontend/src/views/modules/common/createItemsLogic.js
