+++ frontend/src/views/modules/common/createReducer.js
