+++ frontend/src/views/modules/common/fakeServer.js
@@ -1,7 +1,7 @@
 /* Backend 서버를 대신해 임시로 만든 fakeServer
  API 를 정의함
 */
-import { samroads } from "../../data/samroads.json";
+import { samroads } from "../../../data/samroads.json";
 
 function MakeDataGenerator(items) {
   let itemIdx = 0;
