+++ frontend/src/views/modules/friend/state.js
