+++ frontend/src/views/modules/samroad/SAMroadList.js
@@ -1,7 +1,7 @@
 /* Container 가 아닌 presentational component */
 import React from "react";
 import { useSelector } from "react-redux";
-import Post from "../../views/components/post/Post";
+import Post from "../../components/post/Post";
 
 export default function SAMroadList({ samroads }) {
   // keyword 에 따라 post 를 filter
