+++ frontend/src/views/pages/MainPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../components/header/Header";
 import MenuBar from "../components/menubar/MenuBar";
-import SAMroad from "../../modules/samroad/SAMroad";
+import SAMroad from "../modules/samroad/SAMroad";
 import "./Pages.css";
 
 export default function MainPage() {
